trying to fix db

+++ backend/migrations/versions/156b657fe68a_add_missing_columns_to_projects.py
@@ -7,6 +7,7 @@ Create Date: 2025-07-11 16:12:59.007459
 """
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy import inspect
 
 
 # revision identifiers, used by Alembic.
@@ -17,9 +18,20 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column('projects', sa.Column('tariff_id', sa.Integer(), nullable=True))
-    op.add_column('projects', sa.Column('custom_flat_rate', sa.Numeric(precision=10, scale=4), nullable=True))
+    bind = op.get_bind()
+    insp = inspect(bind)
+    cols = [c['name'] for c in insp.get_columns('projects')]
+    if 'tariff_id' not in cols:
+        op.add_column('projects',
+                      sa.Column('tariff_id', sa.Integer(), nullable=True)
+                      )
+    if 'custom_flat_rate' not in cols:
+        op.add_column('projects',
+                      sa.Column('custom_flat_rate', sa.Numeric(precision=10, scale=4), nullable=True)
+                      )
 
 def downgrade():
-    op.drop_column('projects', 'custom_flat_rate')
-    op.drop_column('projects', 'tariff_id')
+    if op.get_bind().dialect.has_table(op.get_bind(), 'projects'):
+        # only drop tariff_id if you really want to roll it back
+        op.drop_column('projects', 'custom_flat_rate')
+        op.drop_column('projects', 'tariff_id')
